// src/app/api/models/flight-offer-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface FlightOfferDto {
  adults?: number;
  children?: number;
  currencyCode?: string;
  departureDate?: string;
  destinationLocationCode?: string;
  max?: number;
  nonStop?: boolean;
  originLocationCode?: string;
  returnDate?: string;
  travelClass?: string;
}